+++ AppVersion.cs
@@ -2,8 +2,8 @@ namespace NOCS
 {
     internal static class AppVersion
     {
-        public const string ReleaseBase = "0.5.4";
+        public const string ReleaseBase = "0.5.6";
         public const string BepInSemVer = ReleaseBase;
-        public const string DisplayVersion = "0.5.4QV";
+        public const string DisplayVersion = "0.5.6QV";
     }
 }